REF: Rename metrics/metrics.py to metrics/_metrics.py
